Move keyboardShortcuts to shared
