+++ web/src/nav/GlobalNavbar.tsx
@@ -8,7 +8,7 @@ import * as GQL from '../../../shared/src/graphql/schema'
 import { PlatformContextProps } from '../../../shared/src/platform/context'
 import { SettingsCascadeProps } from '../../../shared/src/settings/settings'
 import { authRequired } from '../auth'
-import { KeyboardShortcutsProps } from '../keyboardShortcuts/keyboardShortcuts'
+import { KeyboardShortcutsProps } from '../../../shared/src/keyboardShortcuts/keyboardShortcuts'
 import { parseSearchURLQuery, PatternTypeProps } from '../search'
 import { SearchNavbarItem } from '../search/input/SearchNavbarItem'
 import { EventLoggerProps } from '../tracking/eventLogger'
