+++ web/src/nav/NavLinks.test.tsx
@@ -6,7 +6,7 @@ import { setLinkComponent } from '../../../shared/src/components/Link'
 import { ExtensionsControllerProps } from '../../../shared/src/extensions/controller'
 import * as GQL from '../../../shared/src/graphql/schema'
 import { SettingsCascadeProps } from '../../../shared/src/settings/settings'
-import { KeyboardShortcutsProps } from '../keyboardShortcuts/keyboardShortcuts'
+import { KeyboardShortcutsProps } from '../../../shared/src/keyboardShortcuts/keyboardShortcuts'
 import { eventLogger } from '../tracking/eventLogger'
 import { NavLinks } from './NavLinks'
 import { ThemePreference } from '../../../shared/src/theme/theme'
