+++ web/src/nav/NavLinks.tsx
@@ -15,7 +15,7 @@ import {
     KEYBOARD_SHORTCUT_SHOW_COMMAND_PALETTE,
     KEYBOARD_SHORTCUT_SWITCH_THEME,
     KeyboardShortcutsProps,
-} from '../keyboardShortcuts/keyboardShortcuts'
+} from '../../../shared/src/keyboardShortcuts/keyboardShortcuts'
 import { ThemePreferenceProps, ThemeProps } from '../../../shared/src/theme/theme'
 import { EventLoggerProps } from '../tracking/eventLogger'
 import { fetchAllStatusMessages, StatusMessagesNavItem } from './StatusMessagesNavItem'
