+++ web/src/nav/UserNavItem.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 import { ButtonDropdown, DropdownItem, DropdownMenu, DropdownToggle } from 'reactstrap'
 import * as GQL from '../../../shared/src/graphql/schema'
-import { KeyboardShortcut } from '../keyboardShortcuts/keyboardShortcuts'
+import { KeyboardShortcut } from '../../../shared/src/keyboardShortcuts/keyboardShortcuts'
 import { ThemePreference, ThemePreferenceProps, ThemeProps } from '../../../shared/src/theme/theme'
 import { UserAvatar } from '../user/UserAvatar'
